+++ examples/mariadb/outputs.tf
@@ -1,6 +1,1 @@
-output "address" {
-  value = [
-    module.rds_instance_mariadb,
-  ]
-  description = "Output values for the module"
-}
+# Empty