+++ examples/minimum/data.tf
@@ -1,7 +1,18 @@
-data "aws_availability_zones" "available" {
-  state = "available"
+data "aws_vpc" "supporting" {
+  filter {
+    name   = "tag:Name"
+    values = [local.supporting_resources_name]
+  }
 }
 
-data "aws_caller_identity" "current" {}
+data "aws_subnets" "database" {
+  filter {
+    name   = "tag:Name"
+    values = ["${local.supporting_resources_name}.isolated.*"]
+  }
+}
 
-data "aws_region" "current" {}
+data "aws_subnet" "database" {
+  for_each = toset(data.aws_subnets.database.ids)
+  id       = each.value
+}