+++ examples/s3-import/locals.tf
@@ -1,8 +1,14 @@
 locals {
-  name             = "mysql-import-example"
-  db_name          = "randomDB"
-  cidr_block       = "172.16.0.0/16"
-  database_subnets = cidrsubnets(local.cidr_block, 8, 8, 8)
-  environment      = "examples"
-  azs              = flatten(data.aws_availability_zones.available.names)
+  name                      = "mysql-s3-import-example"
+  db_name                   = "examples3db"
+  vpc_id                    = data.aws_vpc.supporting.id
+  database_subnets          = local.database_subnet_id
+  supporting_resources_name = "terraform-aws-rds"
+  database_subnet_id = [
+    for s in data.aws_subnet.database : s.id
+  ]
+  tags = {
+    Name        = local.name
+    Environment = "examples"
+  }
 }