+++ examples/s3-import/main.tf
@@ -1,20 +1,6 @@
 #################################
 # MySQL Import from S3 example
 #################################
-module "vpc" {
-  source               = "git::https://github.com/boldlink/terraform-aws-vpc.git?ref=2.0.3"
-  cidr_block           = local.cidr_block
-  name                 = local.name
-  enable_dns_support   = true
-  enable_dns_hostnames = true
-  account              = data.aws_caller_identity.current.account_id
-  region               = data.aws_region.current.name
-
-  ## database Subnets
-  database_subnets   = local.database_subnets
-  availability_zones = local.azs
-}
-
 resource "random_string" "rds_usr" {
   length  = 5
   special = false
@@ -33,7 +19,7 @@ module "s3_import" {
   engine         = "mysql"
   engine_version = "8.0.28"
   instance_class = "db.t2.small"
-  subnet_ids     = flatten(module.vpc.database_subnet_id)
+  subnet_ids     = local.database_subnets
   port           = 3306
   name           = local.db_name
   username       = random_string.rds_usr.result
@@ -46,27 +32,24 @@ module "s3_import" {
   }
 
   kms_key_id                          = data.aws_kms_alias.rds.target_key_arn
-  environment                         = local.environment
   iam_database_authentication_enabled = true
   multi_az                            = true
   enabled_cloudwatch_logs_exports     = ["general", "error", "slowquery"]
   create_security_group               = true
   create_monitoring_role              = true
   monitoring_interval                 = 30
-  create_option_group                 = true
   deletion_protection                 = false
-  vpc_id                              = module.vpc.id
+  vpc_id                              = local.vpc_id
   assume_role_policy                  = data.aws_iam_policy_document.monitoring.json
   policy_arn                          = "arn:${data.aws_partition.current.partition}:iam::aws:policy/service-role/AmazonRDSEnhancedMonitoringRole"
   major_engine_version                = "8.0"
+  tags                                = local.tags
   options = {
     option = {
       option_name = "MARIADB_AUDIT_PLUGIN"
     }
   }
-  other_tags = {
-    "cost_center" = "random"
-  }
+
   depends_on = [
     aws_s3_bucket.mysql,
     aws_iam_policy.s3_bucket