+++ examples/s3-import/outputs.tf
@@ -1,6 +1,1 @@
-output "s3_import" {
-  value = [
-    module.s3_import,
-  ]
-  description = "Values for resources created by the module."
-}
+# Empty